Update package dependencies and add user routes for signup and login pages
This commit updates the package dependencies by adding "uuid" to the package.json file and updating the package-lock.json and .package-lock.json files. Additionally, new user routes for signup and login pages have been added to the "shortner.js" file in the "routes" directory. However, the actual implementation of these routes is not provided in the given code diff.